storage: Remove dead code from taskManager.prepareForShutdown()
We have a piece of code attempting to stop all tasks listed in the
taskManager dictionary, however since it is a dictionary, the traversed
keys are task ID strings and not Task objects. Hence we have a futile
attempt to check if the key has a stop attribute while we never call
stop() for the actual Task instance that way.

Since this is an ancient code dating back to commit c16678498
which does not really work from its very first day, and nobody expects
it to do so as threadPool is shut without actually waiting for tasks
to stop and finish, we can get rid of it.

Change-Id: I618fac87ad6acc452bbfdde138ccf4303b027fd9
Signed-off-by: Amit Bawer <abawer@redhat.com>