fileSD: Make detect_block_size compatible with Gluster
We used to detect block size by creating __DIRECT_IO_TEST__. This make
it possible that multiple hosts will overwrite the same file, but this
is not an issue since we don't care about the file contents, only about
successful write to the file.

Unfortunately this triggers a bug in Gluster, when creating a file with
O_TRUNC from multiple hosts can corrupt Gluster own metadata. This need
to be fixed in Gluster, but we need to be compatible with existing
Gluster deployments that have this bug.

ioprocess 1.3.0 is implementing now probe_block_size() API compatible
with existing Gluster storage; use this API.

Because of CI caching we must require exact version of
pyhton{2,3}-ioprocess package in the automation packages files. This
will require an update every time ioprocess is updated but we don't have
a better solution now.

Change-Id: Ib8400439f2353d100d7fe56c9ea6250dc43fc726
Bug-Url: https://bugzilla.redhat.com/1753898
Depends-on: https://gerrit.ovirt.org/c/103468/
Signed-off-by: Nir Soffer <nsoffer@redhat.com>