net, nmstate: Remove old default route
Default route network with static IP and gateway stores the
default route in NM. When the configuration of this network
changes from static to DHCP the route is still present
which results in failure of NM becasue we cannot specifiy
default route without gateway.

Second scenario is when the network still remains with static
IP, but default route status is removed. We need to explicitly
remove default route.

Change-Id: I55c0ac92ac189ba74adb35f42168807840607513
Signed-off-by: Ales Musil <amusil@redhat.com>