py3: pylint: Silence catching-non-exception false-positives
Py3-based pylint reports multiple 'catching-non-exception' errors for
'vdsm.gluster.cli' module:

 Catching an exception which doesn't inherit from Exception: _etreeExceptions (catching-non-exception)

The problematic piece of code is the '_etreeExceptions' tuple which
contains 3 exception classes. Since using the tuple to catch multiple
exception classes is valid in py3, and most exceptions listed are
built-in ones, the only possible offender is the 'etree.ParseError'
class. Trying this simple test:

 import xml.etree.ElementTree as etree
 issubclass(etree.ParseError, Exception)

gives us 'True' for both py2 and py3.

This looks like a false-positive, so we're silencing it.

Change-Id: I22a9980c01a897f4d3edffe45fc2db0e00c450fb
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>