tool: Don't invoke systemctl.show in import time
Fix an issue introduced by:

commit a5218e80acc74597fff413b5519c260503611044
Author: Marcin Sobczyk <msobczyk@redhat.com>
Date:   Fri Sep 13 08:32:54 2019 +0200

    tool: Adjust to libvirt's systemd socket activation

When libvirt.FILES is initialized during import time by invoking
systemctl.show(). Running commands during import time is bad idea in
general, and in particular for testing. This call fails when running the
tests in a container without systemd, and there is no way to monkeypatch
the call during import.

Change libvirt.FILES to use conditional fragment for "LIBVIRTD_ARGS",
matching when libvirt does not use socket activation. Check if libvirt
uses socket activation when calling configure(), so it is easy to
simulate both cases in the tests.

Change-Id: I21bf198e31cd915e939302c75a7b483bd8a5b084
Signed-off-by: Nir Soffer <nsoffer@redhat.com>