tests, net, nmstate: let NM manage dummies on nmstate backend
Up to now, we've relied on the network manager configuration to
indicate that the dummy interfaces created in vdsm's network
functional tests are to be managed by NetworkManager, as explained
in [0].

Despite that, the func test framework cleanup is done by the test
framework, in [1]. This patch thus centralizes the configuration of
the dummy interfaces by NetworkManager in the framework, as is
currently done for veth interfaces.

This homogenizes the framework, thus leaving all configuration to be
done the same way, independently of the interface type, via the test
framework.

The network functional tests README file is updated accordingly to
reflect this change.

[0] - https://github.com/oVirt/vdsm/tree/master/tests/network/\
functional#configuring--installing-the-environment
[1] - https://github.com/oVirt/vdsm/blob/master/tests/network/\
nettestlib.py#L263

Change-Id: I51a1e4223bda6f4bf4f064e7ddc8c3e29ad64fb6
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>