tests, net: Fix improper rollback func test asserts
Due to wrong indentation, several asserts have not been executed
at all.

By fixing the asserts on the rollback tests, a bug is exposed when
the OVS switch type is used, since, as can be seen in [0], the
'rollback' option used should be '_inRollback'. The fix to it is
also contained in this patch.

[0] - https://github.com/oVirt/vdsm/blob/\
99cef25660a78390d0b53077c92aab1fb8eac08f/\
lib/vdsm/network/api.py#L204

Change-Id: I331fda8f1960662a1026b662ff0d4f2cee8fd73e
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>