concurrent: Avoid unneeded copies
We used to create a new tuple with all values before submitting the
values to pending queue. This increase the memory usage and the time to
submit the values for no benefit.

Avoid the copy by using pending.qsize() to get the number of values. Use
six.moves.range to avoid another unneeded list when iterating over the
results.

I could not measure significant performance difference with this change.

Change-Id: Ie623542480747507a868a473acd83d150da78764
Bug-Url: https://bugzilla.redhat.com/1598266
Signed-off-by: Nir Soffer <nsoffer@redhat.com>