storage.exception: Add missing attributes
When validating image move, code raising ImageDoesNotExistInSD was
passing template info to the error handler by adding attributes to the
exception instance.

This kind of code is worst practice and causes pylint to complain. Add
the missing attributes to the exception class to make the code more
clear and make pylint happy.

Change-Id: I0678911376d3229cb75bf1e6b3afd208b6a93a35
Bug-Url: https://bugzilla.redhat.com/1755829
Signed-off-by: Nir Soffer <nsoffer@redhat.com>