mount: Include failure details in MountError
Change MountError to inherit from cmdutils.Error, and include the
command, rc, out, and err when raising an error. This allows callers to
log helpful messages and include the info in public exceptions.

Change BlockStorageDomainMasterMountError to include also stderr from
the failing mount or umount command. Having the info in vdsm and engine
logs can make it easier to debug mount issue.

This also fix python 3 compatibility issue when older code tried to
unpack the exception to the arguments. Now we can use cmdutils.Error
attributes.

Change-Id: I3e5d01a495de4852d9465956e0186b399d62fc03
Signed-off-by: Nir Soffer <nsoffer@redhat.com>