ci: Fix yum-builddep usage for build-artifacts
When running build-artifacts sub-stages, just after doing 'make', we
install build dependencies for the RPM. To do that, we're using
'yum-builddep' and pointing to the freshly generated 'vdsm.spec' file.

What we're missing here is lack of 'target_py' variable definition which
defaults to 'py2' in the spec. This makes yum thinking, that we always
want dependencies for py2-based RPMs.

This patch fixes the issue by using 'CI_PYTHON' env variable,
converting its value to 'py2' or 'py3' and passing it as a macro
definition for 'yum-builddep' command.

Change-Id: If5aa59ca96ba1d297a0d2f283f00c3dc805ca2e0
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>