virt: Simplify Graphics device class
Only XML Engine (>=4.2) is supported now.  That means we needn't
support the legacy device class and related legacy functionality
anymore.  The bits from it we still need can be kept in separate
functions and in a simple class, to keep device setup and teardown
working properly, out of the complex device tracking and processing
mechanism.

Change-Id: Ic04c69055417fc70e1b8beafb0264eada065a686
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>