ci: Run tests on el8
This patch enables py3 tests on el8.

We don't have native-el8 CI slaves yet, so we can't use:

 host-distro: same

option anymore. Dropping it would leave us in a state, where the CI
sub-stages could be run on el7 with mocked el8. This causes storage
tests to fail, since they require a more recent kernel version to run
successfully. Thus, we switched to:

 host-distro: newer

The effect is using native-fc30 slaves with el8 mocked. With this option
the fc30 sub-stage is also run on native-fc30 slaves.

Change-Id: I632624da15f848890d38b1ec4e45ffbb0f038d96
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>