storage: return dict values in lvm module also in py3
Replace dict.values() in lvm module when returning them. This results
in different return type in py2 and py3 and also couple of pylint
warnings in py3.

To get list of values in py3, we need to convert returned view
to list. However, in py2 this would result into double copy of the
values (one in py2 dict.values(), which creates copy of the values
and second when creating the list), therefore we use six.itervalues()
and convert it to list.

Change-Id: I7608b2410b88c07fda815c8d652b550165f5218f
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>