automation: Add el8 repos and packages for check-patch
This patch adds the dependency list and the repo file needed to run
'check-patch' sub-stages with el8. Currently some of the packages needed
to run the tests are missing (namely: mom, python3-ioprocess) and some
need to be pulled from Sandro's internal repo . When this is fixed,
we should drop the internal repo and add back the missing packages.

Change-Id: I9c73c6302e2ddb6dfcfd8774481c6c4ef2ec8125
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>