hooks: Improve hooks' environment preparation
Python 3 handles encoding env variables by itself, so no manual encoding
is not needed. For Python 2 we're switching to usage of default system
encoding instead of forcing utf-8.

In our 'tox.ini' file we switch to 'C' locale [1], which happens
to be 'ascii' or 'ANSI_X3.4-1968' (one of ascii's revisions [2])
on some of our slaves. There's no chance that localized characters will
work at all with an ascii locale - the test case is expected to fail in
this situation, so it's been marked with xfail.

[1] https://github.com/oVirt/vdsm/blob/402bde5a94cfc0f253d966e54909a5c225694bc5/tox.ini#L15
[2] https://stackoverflow.com/questions/48743106/whats-ansi-x3-4-1968-encoding

Change-Id: I930fc435ea5498009a9b2003fb37d369073f586d
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>