storage: always use lock when modifying lvm dict
When reloading all LVs, we modify internal LV dictionary, but don't
use lock when doing it, so it can be modified while another thread
is iterating during the LVs dict.

Acquire the lock before we do any changes in the LVs dict. We should
acquire before we check for stale LVs to be sure that the data we will
use later on for modifying the dict is up to date.

Change-Id: Ieee36aec40de2bd51abfc4b214b41ede41db93cf
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>