tests, storage: Add roundtrip test to mailbox_test
Added test covers flow of the HSM-to-SPM extension request message,
the SPM-to-HSM extension reply message and the resulting clearing of
the handled message in HSM outbox.

It also exposes following exception met for storage-py3 testing

  File "vdsm/lib/vdsm/storage/mailbox.py", line 330, in _handleResponses
    MESSAGE_SIZE * "\0" + self._outgoingMail[start +
TypeError: can't concat str to bytes

So it is currently marked with xfail_python3 to be repaired in next patch.

Change-Id: I1bea996e0e53b2dec8734aea58a3c342b94154dd
Signed-off-by: Amit Bawer <abawer@redhat.com>