tests: add test reproducing issue with stale LVs
When iterate over LVs and remove stale LVs, current code use generator,
which use same dict from which we remove stale LVs, which leads to

    E   RuntimeError: dictionary changed size during iteration

It worked on py2 as dict.keys() returned list, while in py3, when
generators is returned, it fails. In previous patch we removed
calling keys() functions and iterate over dict, so now it fails
always and makes this issue more clear.

This patch adds a test which reproduces this issue. The issue itself
will be fixed in subsequent patch.

Change-Id: I9ae726708a96eeb107aee9cffa0340a1de151e8c
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>