storage: replace LVM dict with fresh values
When loading all LVs, create new dict from the obtained values and
replace LVMCache._lvs dict with this one. This ensures we have fresh
values in the cache after calling LVMCache._loadAllLvs() and using this
approach we don't have to deal with removing stale LVs from the dict so
the code for removing stale LVs can be deleted now.

Also rename LVMCahce._reloadAllLvs() to _loadAllLvs() to
better reflect the purpose of this method in its name.

Change-Id: I6253b8080dbe5ac5044cf183879b75792d8922c1
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>