helpers: Fix storage helpers on CentOS 8 build
In CentOS 8, "fallocate" and "managedvolume-helper" fail to run with:

    vdsm.common.cmdutils.Error: Command ['../helpers/fallocate', '4096',
    '/var/tmp/vdsm/test_allocate0/image'] failed with rc=127 out=b''
    err=b'ionice: failed to execute ../helpers/fallocate: No such file
    or directory\n'

This happens because the script using:

    #!/usr/bin/python2

but python2 is not installed on CentOS 8.

Fix by running the scripts using sys.executable:

    sys.exectuable /path/to/script

This is ugly, assuming that the scripts are using python, but it is the
simplest way to have them work on both python 2 and 3, and enable the
CentOS 8 build.

Change-Id: I1f189ddadf3302c7f8a1be6107af9f1cb7858d53
Signed-off-by: Nir Soffer <nsoffer@redhat.com>