integ tests, net: test exploded notation ipv6 addresses
When used against the nmstate backend, it is not possible
to correctly configure logical networks using static ipv6
addresses in the exploded notation format, since nmstate
interprets the same IP address in exploded and compressed
notations to be a different IP address. This leads to failure
when verifying if the desired state was applied.

The supplied test also fails against the init script backend,
but for a different reason: the kernel configuration and the stored
configurations are different, since their IPv6 address formats
differ.

This patch adds an xfailed test that checks this behavior.

Change-Id: I72c99323da783f62446e2017fc20397902eff086
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>