py3: utils: Fix '_parseCmdLine' and 'utils_test'
'utils._parseCmdLine' function is used to read out processes' arguments
from their '/proc/${pid}/cmdline' files. The issue is it tries to split
the read file with a string pattern which doesn't work in py3.

The test that uses the function was also broken on py3 - it compared
the read data (bytes) with string literals (unicode) passed
to 'commands.start'.

Additionally a call to deprecated 'execCmd' function has been replaced
with 'commands.run'.

Change-Id: Ie2dee2e6d9766f26ca6eb15339d89d6f22583370
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>