py3: tests: Port 'bridge_test' to py3
"imp" module is deprecated and it's usage should be removed.

This patch redefines the creation of dynamic "vdsm.API" module in terms
of newer "importlib" module. Unfortunately this module is very limited
on py2, so we're dropping py2-compatibility for these tests.

Python 3 also doesn't support negative values for the "level" argument
passed to "__import__" [1] (which means an attempt of both absolute
and relative import on py2), so we're switching to the default value
of 0 (which means absolute imports only) since it works for us.

[1] https://docs.python.org/3/library/functions.html?highlight=__import__#__import__

Change-Id: I40ad4ce86f520dd7de9d9426d693eec788ebbfa5
Bug-Url: https://bugzilla.redhat.com/1663661
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>