stomp: Fix improper frame terminator search
A bug has been introduced in [1] - since indexing bytes yields slices
in py2, and ints in py3, we've forced a slicing operation on the buffer
when searching for frame terminator to have a consistent result accross
both interpreter versions. Unfortunately in case where the buffer holds
more data after the null byte, the comparison fails causing the parser
to raise an exception. This patch fixes this issue by replacing slicing
with indexing and using different objects for comparison on py2 and py3.

Additionally, a simple test for parsing multiple frames that covers
the described case has been added for the 'stomp.Parser' class.

[1] https://gerrit.ovirt.org/#/c/101114/

Change-Id: I2368ffaa1ec399b570e2dd6ddc1bab5e0825127a
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>