py3: tests: Skip 'unicode_test' tests on py3
Tests inside 'unicode_test' module are used to check the procedures that
make utf-8 the default encoding for py2 in
'static/usr/share/vdsm/sitecustomize.py'. They are not relevant and
can't really work on py3 though due to string vs byte literals issues.

We can't move this test module to tox since this would skip
the aforementioned procedures. The only nice way to handle this test
module is to simply skip the tests inside for py3.

Change-Id: I189cb88b1f09ba69e358caeaca129e26fef817d0
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>