tests: Remove black magic from makefile
There are no more py3-blacklisted test modules. We've also switched from
usage of py27/py36/py37 suffixes to py2/py3 ones in the makefiles.
There's no reason to keep the black magic that was needed before. Let's
simplify things to make it more easy to switch to py3 completely in the
future.

Change-Id: Ib79a5d38bc1e73e398a7457fa6991d6284c51ff5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>