drop vdsm_hook/promisc
The `promisc` hook was created in order to allow Engine administrators
configure a VM that capture all traffic of other VMs. This functionality
is available by the Port Mirroring option of a vNIC profile.

The hook has not been used or tested for many years. It is time to drop
its code.

Change-Id: I9d05f4c2ee2d84861680b6f13e3136ec10431614
Signed-off-by: Dan Kenigsberg <danken@redhat.com>