vdsm-tool: Drop support for 4.16 upgrade fix
This patch drops a fix that was about importing 'vdsm.tool' from wrong
location when upgrading from 4.16 version. This is not relevant anymore.

After converting imp-based import of 'vdsm.tool' to a regular one,
pylint found an issue with an exception handling code:

 E:215,11: Bad except clauses order (UsageError is an ancestor class of ExtraArgsError) (bad-except-order)

Thus, the order of exception catching has also been changed.

Change-Id: Ifda592d41bdde6688854e40300051fc7481624dd
Bug-Url: https://bugzilla.redhat.com/1663661
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>