threadPool: Remove online resize feature
We did not use this feature in the last 10 years, and we don't plan to
use it.

Since we don't need to support online resizing, inline the code to start
the workers threads in __init__(), and replace the call to resize the
thread count to 0 with single del statement.

Finally since we start the thread only in __init__, the _count instance
variable is not needed now.

Change-Id: Ia340e0fa49a5ccd85a51cc664c5f7a18e28ad0fb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>