storagetestlib: Move Callable timeout to constructor
Callable.__call__() accepts a timeout argument, which can be used to
limit the time a callable will hang during the tests. This is a good
idea (although unused yet), but we do not control __call__() arguments,
so better move it to __init__().

Replace "hang" argument with "hang_timeout". The default value is 0
keeping the previous behaviour. To create a hanging callback create it
with the maximum timeout instead of True.

Change-Id: I245b99e3e10b30e7dd01fc55383739578589e669
Signed-off-by: Nir Soffer <nsoffer@redhat.com>