storage, mailbox: Fix free message slot lookup in HSM mail monitor
Current logic initializes freeSlot index as False and checks for its vacancy with
a negation condition. This creates a bug falsly ignoring zero index as a possible slot,
not allowing to utilize it for an added message.

This issue was reflected in a raised exception of full active messages list for the
priorly added test_fill_slots(), so the xfail marker is removed here for this test.

Change-Id: I294d578f9613279812b178675b49249b8d0562fe
Signed-off-by: Amit Bawer <abawer@redhat.com>