tests: Restore 'jsonRpcTests' module
'jsonRpcTests' module has been hiding in the shadows for a long time -
it's been moved from 'tests/' directory to 'tests/integration' in [1],
but at the same time accidentally removed from a list of test modules
to be run. Apart from usage of no longer existing 'stomp.Disconnected'
exception [2] and some minor linting issues it seems to be doing fine
though. Let's remove the obsolete tests and restore the rest
of the module.

[1] https://github.com/oVirt/vdsm/commit/0d591d352a81bddb3dc4e8b444fb1098900f057c
[2] https://github.com/oVirt/vdsm/commit/e62cc7c4269a5a93d7c533bb740ca46ad3c0926a

Change-Id: Ia1b3b9e430492d36392fc26678ec92481fcbbdc0
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>