tests: Remove 'DEAFAULT_SSL_CONTEXT' dependency in 'constructAcceptor'
In an effort to drop 'vdsm-tests' package, we need to get rid
of build-time generated SSL keys and certificates created
by 'tests/makecert.sh' script and the script itself.

This patch changes the 'constructAcceptor' helper from
'integration.jsonRpcHelper' module.

This method used to take an 'ssl' argument which decided whether
or not to pass an SSL context to 'MultiProtocolAcceptor'. When called
with 'ssl=True', the global 'DEAFAULT_SSL_CONTEXT' was used.

The 'ssl' parameter has been changed to 'ssl_ctx', which allows to pass
the context directly. This removes the dependency
to 'DEAFAULT_SSL_CONTEXT'.

'stomp_test' module has been the only client of this helper, so it's
been adapted to the new way of working.

Change-Id: I8d57bb3c39cd3d0b92c96b29d57cc83a38cd501c
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>