py3: tests: Move 'client_test' module to tox
In an effort to clean up 'tests' directory, we're moving
'client_test' module to 'tests/lib/yajsonrpc' dir and run it with tox.

The tests inside the module were skipped for py3, but it turns out that
they're now working fine with this interpreter.

The module has also been renamed to more accurately reflect the tested
codebase.

Change-Id: I8fc45ac7a87107e32d6cc6821d5954dc3c9d3fbc
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>