tests: tests: Remove 'DEAFAULT_SSL_CONTEXT' dependency in 'constructClient'
In an effort to drop 'vdsm-tests' package, we need to get rid
of build-time generated SSL keys and certificates created
by 'tests/makecert.sh' script and the script itself.

This patch changes the 'constructClient' helper from
'integration.jsonRpcHelper' module.

This method used to take an 'ssl' argument which decided whether
or not to pass an SSL context to 'constructAcceptor' and for socket
creation. When called with 'ssl=True', the global 'DEAFAULT_SSL_CONTEXT'
was used.

The 'ssl' parameter has been changed to 'ssl_ctx', which allows to pass
the context directly. This removes the dependency
to 'DEAFAULT_SSL_CONTEXT'.

'jsonrpcserver_test' and 'stomprpcclient_test' modules were clients of
this helper, so they've been adapted to the changes.

Change-Id: I1d2e74694fa0e82a72b84b92171f16d7695659d3
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>