build: Drop 'vdsm-tests' package
'vdsm-tests' package has been broken an unused for a long time - let's
drop it.

Since 'tests/run_tests_local.sh' file is not part of dist package
anymore, it is not built along with the 'all' target. 'check-*' targets
from 'tests/Makefile.am' have an explicit dependency to this script,
so they will do fine, but a small adjustment to 'contrib/shell_helper'
has been added to make sure the script is generated before trying to run
the tests.

Change-Id: Ib531b68a476d9b4dd4cd942734667424b2ac4f1b
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>