storage, mailbox: Replace numBytes parameter by CHEKSUM_BYTES constant for checksum
We always use the 4 bytes constant for checksum passed length, so no need to pass
it explicitly.

The additional result trimming into 32 bit is kept in a more readable format
for sake of robustness in case we might use the checksum over input data summing
up to more than a 4 bytes value.

Change-Id: I7425bed3c15b1708c63936c4f2ec478dfd440cef
Signed-off-by: Amit Bawer <abawer@redhat.com>