network: Avoid dracut overwriting ifcfg files
If dracut detects that networking might be required to boot,
dracut will overwrite ifcfg files. This might lead to
unexpected network configuration after boot.
To prevent this situation, vdsm instruct dracut to avoid
touching the ifcfg files.

This becomes relevant now, because
https://gerrit.ovirt.org/#/c/103009/ introduced a dependency
to clevis-dracut, which triggers because of bug 1762028
the activation of network and overwriting ifcfg files on boot.

Bug-Url: https://bugzilla.redhat.com/1760262
Change-Id: I5e38117a116cc7b29b9092d154eaeae8a6c2a1b4
Signed-off-by: Dominik Holler <dholler@redhat.com>