virt: change live snapshot to async
This patch will allow the operation of live snapshot
to be async. The live memory dump was the only
operation that was synchronized.

Now, snapshotCreateXML will be executed on a
separate thread. Making the live memory dump to be
async operation.

The job UUID given from the engine will be used to
add as a job for the VM until it's done and let the engine
monitor the job.

The snapshot job will use Virt jobs mechanism in order to
know if the memory dump is active while the engine will
keep polling for the job.

A follow-up patch will handle the recovery of the job.

Change-Id: I202ce346786815d3104475e5f61c235a10528ea1
Bug-Url: https://bugzilla.redhat.com/1749284
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>