vdsmd: change KillMode back to control-group
We've changed KillMode to mixed in

commit a0b9f2c657f2fd5b32ee613ef28811c3e36f6365
Author: Yeela Kaplan <ykaplan@redhat.com>
Date:   Tue Aug 18 18:04:03 2015 +0300

    service: change vdsm KillMode to mixed

This doesn't seems to be correct, as child processes are killed
suddenly with SIGKILL, giving them no chance to terminate properly.

Switch back to default control-group, when all processes obtain SIGTERM
first and are killed by SIGKILL after timeout if they haven't terminated
yet.

Change-Id: I9c19fc64e5ed80ef3d9d64ba63676b798031303c
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>