net: Do not delete bridge when slave is down
When performing an ifdown on an unenslaved nic, which it's ifcfg
file contains a bridge entry, the bridge is wrongly removed.
This is a new behavior on el8, and does not reproduce on el7.

When detecting a nic that is to be enslaved to a nic, initialize
it's configuration through iproute2 instead of ifdown.

Change-Id: I71af62605e63f914ca8af4abf1e5b1b1de17cfc1
Signed-off-by: Bell Levin <blevin@redhat.com>