lvm: disable event activation
lvm2-2.03 has, besides other changes, removed support for lvmetad,
which is not used anymore and introduced event activation, which
is enabled by default. To prevent unwanted lvm activation of newly
added devices, which can lead to data corruption, disable event
activation. Also remove lvmetad configuration, which is not needed
anymore.

As we still support Fedora 30, which use lvm2-2.02, we still have to
support old lvmetad configuration. As lvm ignores options which it
doesn't recognize, we can keep all configuration in the same file and
use one config for all versions.

If lvm finds unknow option, it ignores it and prints a warning for it.
The warnings can be disabled by setting config/checks=0 in lvm.conf.

The warnings are printed out only on Fedora 30/CentOS 7, on RHEL8
local/use_lvmetad seems to be silently ignored even when
config/checks=1.

Change-Id: I64191e329b43b0dc72d599dd0e31ae25638d22a6
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>