tool: add TODO for removing lvmetad from lvm configurator
lvmetad is not suppored since lvm2-2.03 which is present in RHEL8 and
Fedora 31. We still support Fedora 30, so we still have to keep
disabling of lvmetad, but once we stop supporting it, it should be
removed. Addd TODO to appropriate parts of lvm configurator to remove
it once we don't support Fedora 30.

Running this code on Fedora 31 or RHEL8 seems to be harmless, as
systemd masks the service no matter if it exists or not, but it
would be nice to skip this configuration if not needed. This will be
done in subsequent patch.

Change-Id: Ic38bebdcf01793fc971cee48ac8b9576f0dde952
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>