lvm: add TODO for removing read_only and locking_type
locking_type=1 is the default in RHEL-8 and Fedora 31 but
in RHEL-7 and Fedora < 31 locking_type is still needed.

locking_type=4 isn't needed in RHEL-8 since LVM doesn't
modifying the metadata from read commands anymore bug -
https://bugzilla.redhat.com/1553133.

In RHEL-8 and Fedora > 31 LVM will look at the deprecated
locking_type value, and will translate it to an equivalent
command line option and just print a warning message and
ignore it.

This patch adds a TODO for removing the read_only attribute
and changing the log level of the failed validation for the
locking_type in the HSM.

Change-Id: I79aa0f9ae53e722bd539f9d26b226fef47b563c2
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>