image: Remove unused Image.sparsify()
Image.sparsify() was added in 3.6 as a GSOC project, but was never used
by engine or other supported client. In 4.1 we added
SDM.sparsify_volume() which is used by engine, but we kept the old
sparsify in case engine will want to use it.

Moving to RHEL 8 is an opportunity to drop the dead code.

Change-Id: Ic4f1d0a95f4873c4e9cd4f3de4394e33db25372e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>