v2v: GetStats response message fails in Json
Importing VMWare VM to an EL8.1 Host fails to import
due to a binary value that was not being converted
to a string. This caused the Host.getStats response
message to fail at the Json layer.

Adding string conversion to the description field
mitigated the failure. The issue is isolated to
the current master 4.4 branch.

Change-Id: I6994078ecd004cd33f784c290583018957c9ab84
Bug-Url: https://bugzilla.redhat.com/1770889
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>