spec: Fix 'hook-httpsisoboot' dependency
'hook-httpsisoboot' hook was introduced in [1] to allow using
https-served ISOs without the need to upload them to storage domains.
Since this feature was introduced to qemu-kvm-rhev somewhere in el7
lifecycle, the spec required specific version of this package.
For el8 we can use it right away and given that 'vdsm' package
requires newer version of 'qemu-kvm', we can drop this requirement
completely.

[1] https://github.com/oVirt/vdsm/commit/59534fdf4e23d6719aa81f832de39ba37becf25eq

Change-Id: I10078f03ef50fcd0af1cfafc5f5211da5f6d9e15
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>