net, nmstate: update network backend
Must add the nmstate mock to the netinfo_test, since 2 unit tests
were failing.

To make the integration tests work, we also need to disable nmstate
for integration tests, since they rely on dummy interfaces, which
when using the nmstate backend, are configured w/ NetworkManager,
which is not available on that environment.

Change-Id: I524c141126c32dc1a0da0d270dbc4b18685b0f0e
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>