tests: add dm mapping fixture
Add a fixture which creates a test device mapper mapping. This mapping
uses zero target, which is supposed to be used for device mapper tests
and acts like /dev/zero. For now, the size of the divice is fixed to
1 GiB.

The tests using this fixture need to run with root privileges as
dmsetup utility needs root privileges. The fixture will be used in
subsequent commit for real device mapper tests.

Change-Id: Idcaf05d472c8852d8c8ea56111922a53555f672f
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>