tests: add test for removing dm mapping
Add test for removing device mapper mapping. The test uses real mapping
created by zero_dm_device fixture introduced in previous commit.

The device mapper doesn't work properly on Docker, it hands during
creation of the zero target mapping, waiting for a semaphore:

    ioctl(3, DM_DEV_CREATE, 0x55eda20de290) = 0
    ioctl(3, DM_TABLE_LOAD, 0x55eda20de270) = 0
    ioctl(3, DM_DEV_SUSPEND, 0x55eda20de170) = 0
    semget(0xd4de04a, 1, 0)                 = 0
    semctl(0, 0, GETVAL, 0xffffffff)        = 2
    semop(0, [{0, -1, IPC_NOWAIT}], 1)      = 0
    semop(0, [{0, 0, 0}], 1

The tests times out in the container, so we have to skip this test on
oVirt CI and Travis.

Change-Id: I9bdd6c6307cd41607b0497e42147e581493ed9ac
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>