spec: Add a more meaningful changelog
The current %changelog entry in vdsm.spec.in is not meaningful in any
way.  Let's add a real ChangeLog to /usr/share/vdsm/doc and replace
spec %changelog content with a reference to it.  This is better than
including the whole ChangeLog, which is > 4 MB, in each of the built
rpm files.

We still keep some date etc. in %changelog to make rpm happy.

Change-Id: Id9640eb5214a8b51c45bf5eb7a9dcbd54747e74e
Backport-To: 4.3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>