net, tests, nmstate: fixes the rollback functests
The rollback tests try to rollback the network state when ever
the connectivity check fails.

When the setupNetworks command fails - which it does in this
scenario since the connectivity check is not successful - the
failed transaction is replayed, thus removing all the elements that
might have leaked in the failed transaction.

Since the configuration on the setup_nmstate function never
updated the configuration, the 'delete' transaction fails since
most of the objects being deleted do not exist.

Saving the configuration before attempting the connectivity check
fixes this.

This happens because vdsm's nmstate wrapper uses the RunningConfig
object instead of using the KernelConfig object.

Change-Id: I5df5d9b5ee1f0498eaacd1fb2281d34d4661855f
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>