net: Add exception handling for double mirroring
Having a vm with multiple interfaces, each has a vnic profile with
port mirroring enabled on it - the port mirroring will be
enabled multiple times on the bridge.

On a new version of the kernel, a different exception is thrown
and thus is not handled in our code.

Add handling of the new exception.

Change-Id: I1c397f361b2abd5b00871208094b782aba0c46fe
Bug-Url: https://bugzilla.redhat.com/1765018
Signed-off-by: Bell Levin <blevin@redhat.com>