tool: Add 'show-default-config' command
Before 4.2, it was possible to print vdsm's configuration
file by running 'vdsm/common/config.py' as a script:

 python /usr/lib/python2.7/site-packages/vdsm/common/config.py

This is not possible anymore due to clash between 'vdsm.common.glob'
module and 'glob' module that comes with standard library.

This patch adds a 'show-default-config' command to 'vdsm-tool' that
prints out the default configuration for vdsm.

Additionally, 'config.py' module can still be used to print out the
default config with this syntax:

 python3 -m vdsm.common.config

Change-Id: I49443a63793e4d589e72031c596b1d0d54064460
Bug-Url: https://bugzilla.redhat.com/1529344
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>