supervdsm: Cleaner shutdown procedure
Supervdsmd has an issue that occurs randomly during shutdown:

 systemd[1]: Stopping Auxiliary vdsm service for running helper functions as root...
 daemonAdapter[12446]: Traceback (most recent call last):
 daemonAdapter[12446]:   File "/usr/lib64/python3.6/multiprocessing/util.py", line 262, in _run_finalizers
 daemonAdapter[12446]:     finalizer()
 daemonAdapter[12446]:   File "/usr/lib64/python3.6/multiprocessing/util.py", line 186, in __call__
 daemonAdapter[12446]:     res = self._callback(*self._args, **self._kwargs)
 daemonAdapter[12446]: FileNotFoundError: [Errno 2] No such file or directory: '/var/run/vdsm/svdsm.sock'

This exception causes systemd to actually restart the service instead
of shutting it down.

In cleanup stage of supervdsm's shutdown procedure we delete the
socket file if it exists, but multiprocessing facilities are doing that
by themselves when shut down properly. The reason we experience this
error randomly is most probably the ugly way we treat supervdsm's
serving daemonic thread.

This patch adds a more polite way of shutting down - we make
a connection to the server, call the 'server.shutdown' method
and join on the serving thread.

After implementing this change, the 'FileNotFoundError' exception occurs
100% of the time, meaning we give multiprocessing facilities a way
to shut down properly. We therefore don't need to delete the socket file
by ourselves anymore.

Bug-Url: https://bugzilla.redhat.com/1778638
Change-Id: I7d63226ac0cac9d917f6c9ede939a0ad8dda3c7a
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>