resourcemanager: Raise internal ResourceDoesNotExist for an unfound resource
The legacy behavior was to hide the KeyError raised by
ResourceManager.acquireResource() flow in Owner.acquire() error handling[1],
not exposing the exception or a False return value to the API calls.

This change adds the exception alert in case of attempting to acquire
an unknown resource, causing the Owner.acquire() call to raise in case
the raiseonfailure flag is set, and returning False for such case as well.

[1] https://github.com/oVirt/vdsm/blob/8e53f0f22f3b52f39f4b47d273c279c9b27f1156/vdsm/storage/resourceManager.py#L687

Change-Id: I77c4b3b80bd76941610511888a229267ff6b2f30
Signed-off-by: Amit Bawer <abawer@redhat.com>