virt: Use libvirt migration parameter for host name validation
Encrypted migrations check that the destination certificate matches
the migration destination.  However, this is not always true, for
instance when using a migration network IP address as destination, not
present in the destination host certificate.  We had to implement a
workaround that always uses the standard host name or IP address as
the encrypted migration destination, ignoring migration networks or
other settings.  Otherwise encrypted migrations would fail completely
in such situations.

libvirt provides a new migration parameter,
VIR_MIGRATE_PARAM_TLS_DESTINATION, now that can be used to set the
expected host name in the destination certificate.  Let's use it and
remove the former workaround.

Change-Id: Idccd28613799e40580c39e7fbd002c9a768d2ae7
Bug-Url: https://bugzilla.redhat.com/1739557
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>