net, tests: add dynamic_ip_switch_to_static without DHCP server
Add functional tests for switching from dynamic IP to static IP
without a running dhcp server.

Dhcp without a server is usually problematic when nmstate is
used, which warrants these new tests.

When the DHCP server is *not* started, there won't be any IP
addresses, and as such, the netfunctestslib assertDHCPv[4|6]
and assertRoutesIPv[4|6] have to be updated, adding the ability
to skip IP address checking.

Change-Id: I2b849e105a1e1e458740b64338d3ec44c250fc13
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>