mpathhealth, udev: Remove device name resolution from monitored paths
Handled scenario:

1. Getting path status from "dmsetup status"
   in devicemapper.multipath_status()

2. While processing results, device "8:11" is removed from the system.

3. Trying to resolve the name of the device fails,
   returning "8:11" silently.

4. "8:11" is injected by mistake into multipath health monitor.

For a four paths setup the wrong report would seem like this:

    "valid_paths": 2
    "failed_paths": ["8:11", "sda", "sdb"]

Since we will never get an event for "8:11", we will always report it
in the failed paths until vdsm is restarted.

Dropping the device name resolution has no effect on Vdsm and Engine
operation, as Engine only displays the failing hosts with their relevant
DM_UUIDs and uses the number of failing device paths as an indicator,
regardless of their actual names.

Change-Id: I08bdfb95c2a61a18c037627cbf7473006b9c2452
Bug-Url: https://bugzilla.redhat.com/1651772
Signed-off-by: Amit Bawer <abawer@redhat.com>