net, CI: Add nmstate func tests to check-patch
Preforming network func tests on regular basis will
improve the development of nmstate and help not
introduce new bugs to existing nmstate tests.

The nmstate test are still a little flaky on CI, and sometimes
fail, thus returning as success whether the tests passed or not.

the tests will be run only if some network files were
changed, and do not take much extra time since it is based
on a container.

Change-Id: Ic440e777c68ad2662ad8c68e90caad60c1491a17
Signed-off-by: Bell Levin <blevin@redhat.com>