backup: create checkpoint when starting full vm backup
The first step towards incremental backup is to create
a checkpoint when the backup begins.
The checkpoint will mark a point in time that instructs
Libvirt and Qemu to track the diffs from this checkpoint
to the next one in the chain.

When the user starts a backup, a new checkpoint is
added automatically.

The checkpoint XML structure is as follows:

<domaincheckpoint>
    <name>uniqe_id</name>
    <description>checkpoint for backup 'backup_id'</description>
    <parent>
        <name>parent_id</name>
    </parent>
    <disks>
        <disk name='vda' checkpoint='bitmap'/>
        <disk name='vdb' checkpoint='no'/>
    </disks>
</domaincheckpoint>

A checkpoint is supported for Qcow2 disks only. Therefore,
for raw disks, the checkpoint attribute will be set to 'no'
instead of 'bitmap' according to the engine request.
This will instruct Libvirt to skip the checkpoint creation for
those disks.

Change-Id: Id3a953f5f1e51fe9a46cb2c13f31533e852705a5
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>